[#11] Separate class and instance methods
This commit splits the Familia module's methods into two distinct files:
class_methods.rb and instance_methods.rb. Key changes include:

- Moving ClassMethods to class_methods.rb
- Creating a new file instance_methods.rb for InstanceMethods
- Removing unused methods and redundant code
- Updating the field method implementation in ClassMethods
- Adding debug logging for object registration

This reorganization improves code organization and maintainability.
